Refactor each panel into its own component
Things are diverging quickly enough for us to benefit from this refactor.

* MainView now loads panels without understanding any of them
* I've moved AdminPanel into Screens, and put the new Panels in Screens, because it feels like they're realistically more like screens than like Components (especially given the relative lack of other screens in the whole app)
* I've tried to keep the refactoring as basic as possible.  No style fixes or other fixes: literally ported the functionality.
