+++ src/Screens/MainView.tsx
@@ -1,234 +1,35 @@
 import React from "react";
 import "react-tabs/style/react-tabs.css";
-import Button from "../Components/Button";
-import DataTable from "../Components/DataTable";
-import ImageRow from "../Components/ImageRow";
-import LabelTextInput from "../Components/LabelTextInput";
-import LabelWrapper from "../Components/LabelWrapper";
-import TextItem from "../Components/TextItem";
-import {
-  ClaimEntry,
-  ClaimTask,
-  ReimbursementTask,
-  Task,
-  tasksForRole,
-  UserRole,
-  userRoles
-} from "../store/corestore";
+import { UserRole, userRoles } from "../store/corestore";
 import MainChrome from "./MainChrome";
 import "./MainView.css";
 import { Tab, Tabs, TabList, TabPanel } from "react-tabs";
 import "react-tabs/style/react-tabs.css";
 import "./MainView.css";
-import TaskList from "../Components/TaskList";
-import AdminPanel from "../Components/AdminPanel";
+import AuditorPanel from "./AuditorPanel";
+import PayorPanel from "./PayorPanel";
+import AdminPanel from "./AdminPanel";
 
 type Props = {};
 type State = {
-  notes: string;
   roles: UserRole[];
-  tasksForRoles: Task[][];
-  selectedTaskIndex: number | null;
+};
+
+const RoleToPanelMap = {
+  Auditor: AuditorPanel,
+  Payor: PayorPanel,
+  Admin: AdminPanel
 };
 
 class MainView extends React.Component<Props, State> {
-  state = {
-    notes: "",
-    roles: [],
-    selectedTaskIndex: -1,
-    tasksForRoles: []
+  state: State = {
+    roles: []
   };
 
   async componentDidMount() {
     const roles = await userRoles();
-    const tasksForRoles = await Promise.all(roles.map(r => tasksForRole(r)));
-    this.setState({ roles, tasksForRoles });
+    this.setState({ roles });
   }
-
-  _renderTaskListClaim = (task: Task, isSelected: boolean) => {
-    const claim = task as ClaimTask;
-    const previewName =
-      "mainview_task_preview" + (isSelected ? " selected" : "");
-    const claimAmounts = claim.entries.map(entry => {
-      return entry.claimedCost;
-    });
-    const claimsTotal = claimAmounts.reduce(
-      (sum, claimedCost) => sum + claimedCost
-    );
-    return (
-      <div className={previewName}>
-        <div className="mainview_preview_header">
-          <span>{claim.site.name}</span>
-          <span>{claim.entries.length} Entries</span>
-        </div>
-        <div>{"Claims to Review: " + claim.entries.length}</div>
-        <div>{"Total Reimbursement: " + claimsTotal.toFixed(2) + " KSh"}</div>
-      </div>
-    );
-  };
-
-  _onNotesChanged = (notes: string) => {
-    this.setState({ notes });
-  };
-
-  _onApprove() {}
-
-  _onDecline() {
-    this.setState({ notes: "" });
-  }
-
-  _extractImageURLs = (claim: ClaimEntry) => {
-    const claimImageURLs: string[] = [];
-    if (!!claim.photoIDUri) {
-      claimImageURLs.push(claim.photoIDUri);
-    }
-    if (!!claim.photoMedBatchUri) {
-      claimImageURLs.push(claim.photoMedBatchUri);
-    }
-    if (!!claim.photoMedUri) {
-      claimImageURLs.push(claim.photoMedUri);
-    }
-    return claimImageURLs;
-  };
-
-  _renderClaimEntryDetails = (entry: ClaimEntry) => {
-    let patientProps = [];
-    if (!!entry.patientAge) patientProps.push(entry.patientAge);
-    if (!!entry.patientSex && entry.patientSex!.length > 0)
-      patientProps.push(entry.patientSex);
-    const patientInfo =
-      patientProps.length > 0 ? `(${patientProps.join(", ")})` : "";
-
-    return (
-      <LabelWrapper>
-        <TextItem
-          data={{ Date: new Date(entry.timestamp).toLocaleDateString() }}
-        />
-        <TextItem
-          data={{
-            Patient: `${entry.patientFirstName} ${entry.patientLastName} ${patientInfo}`
-          }}
-        />
-        <TextItem data={{ Item: entry.item }} />
-        <ImageRow imageURLs={this._extractImageURLs(entry)} />
-      </LabelWrapper>
-    );
-  }
-
-  _renderClaimDetails = (task: Task) => {
-    const claim = task as ClaimTask;
-
-    return (
-      <LabelWrapper label="DETAILS VIEW">
-        <TextItem data={{ Pharmacy: claim.site.name }} />
-        {claim.entries.map(this._renderClaimEntryDetails)}
-        <LabelTextInput onTextChange={this._onNotesChanged} label={"Notes"} />
-        <div className="mainview_button_row">
-          <Button label="Decline" onClick={this._onDecline} />
-          <Button label="Approve" onClick={this._onApprove} />
-        </div>
-      </LabelWrapper>
-    );
-  };
-
-  _renderReimbursementDetails = (task: Task) => {
-    const reimbursement = task as ReimbursementTask;
-    const claim = reimbursement.claim as ClaimTask;
-    if (!reimbursement || !claim || !claim.entries) {
-      return null;
-    }
-    const claimAmounts = claim.entries.map(entry => {
-      return entry.claimedCost;
-    });
-    const claimsTotal = claimAmounts.reduce(
-      (sum, claimedCost) => sum + claimedCost
-    );
-
-    let cleanedData: any[] = [];
-    claim.entries.forEach((entry: ClaimEntry) => {
-      let row: any = {};
-      row["Patient"] = `${entry.patientFirstName} ${entry.patientLastName}`;
-      row["Item"] = entry.item;
-      row["Reimbursement"] = entry.claimedCost;
-      cleanedData.push(row);
-    });
-
-    return (
-      <LabelWrapper label="DETAILS VIEW">
-        <TextItem data={{ Pharmacy: claim.site.name }} />
-        {!!claim.site.phone && (
-          <TextItem data={{ Phone: claim.site.phone }} />
-        )}
-        <TextItem
-          data={{ "Total Reimbursement": claimsTotal.toString() + "KSh" }}
-        />
-        <DataTable data={cleanedData} />
-        <LabelTextInput onTextChange={this._onNotesChanged} label="Notes" />
-        <div className="mainview_button_row">
-          <Button
-            label="Payment Complete"
-            onClick={() => console.log("Payment complete")}
-          />
-        </div>
-      </LabelWrapper>
-    );
-  };
-
-  _renderTaskListReimbursement = (task: Task, isSelected: boolean) => {
-    const reimbursement = task as ReimbursementTask;
-    const claim = reimbursement.claim as ClaimTask;
-    const previewName =
-      "mainview_task_preview" + (isSelected ? " selected" : "");
-    const claimAmounts = claim.entries.map(entry => {
-      return entry.claimedCost;
-    });
-    const claimsTotal = claimAmounts.reduce(
-      (sum, claimedCost) => sum + claimedCost
-    );
-    return (
-      <div className={previewName}>
-        <div className="mainview_preview_header">
-          <span>{claim.site.name}</span>
-        </div>
-        <div>{"Total Reimbursement: " + claimsTotal + " KSh"}</div>
-      </div>
-    );
-  };
-
-  _renderRolePane(index: number) {
-    const renderer =
-      this.state.roles[index] === UserRole.AUDITOR
-        ? this._renderTaskListClaim
-        : this._renderTaskListReimbursement;
-
-    return (
-      <TaskList
-        onSelect={this._onTaskSelect}
-        tasks={this.state.tasksForRoles[index]}
-        renderItem={renderer}
-        className="mainview_tasklist"
-      />
-    );
-  }
-
-  _renderDetailsPane(index: number) {
-    const renderer =
-      this.state.roles[index] === UserRole.AUDITOR
-        ? this._renderClaimDetails
-        : this._renderReimbursementDetails;
-    return (
-      <div key={index} style={{ width: "100%" }}>
-        {this.state.selectedTaskIndex >= 0 &&
-          renderer(
-            this.state.tasksForRoles[index][this.state.selectedTaskIndex]
-          )}
-      </div>
-    );
-  }
-
-  _onTaskSelect = (index: number) => {
-    this.setState({ selectedTaskIndex: index });
-  };
 
   _renderBody() {
     if (!this.state.roles.length) {
@@ -242,20 +43,8 @@ class MainView extends React.Component<Props, State> {
             <Tab key={r}>{r}</Tab>
           ))}
         </TabList>
-        {this.state.roles.map((_, index) => {
-          const adminIndex = this.state.roles.findIndex(
-            r => r === UserRole.ADMIN
-          );
-          const panelContent =
-            adminIndex === index ? (
-              <AdminPanel />
-            ) : (
-              <div>
-                {this._renderRolePane(index)}
-                {this._renderDetailsPane(index)}
-              </div>
-            );
-
+        {this.state.roles.map((role, index) => {
+          const PanelTag = RoleToPanelMap[role];
           return (
             <TabPanel
               key={"tab_" + index}
@@ -264,7 +53,7 @@ class MainView extends React.Component<Props, State> {
                 flexDirection: "row"
               }}
             >
-              {panelContent}
+              <PanelTag />
             </TabPanel>
           );
         })}
